MUI Tests - Backend: - Divided tests between API-Tests, Workgroup-API-Tests - Refactored shared test code (modularized) - /env-info now moved into WorkgroupAPI, routes updated everywhere - 404 API handler moved outside - User detection tests added for GCP / Other-IAP / (Basic|Unsecured)
